+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @trigger.dev/react
 
+## 3.0.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [c702d6a9c]
+- Updated dependencies [b271742dc]
+- Updated dependencies [9af2570da]
+  - @trigger.dev/core@3.0.0-beta.3
+
 ## 3.0.0-beta.2
 
 ### Patch Changes